Move subscription badge
